Renamed the Currency files to be Money to match the types.

+++ UnitTests/Components/Currency/CurrencyFormatterTests.swift
@@ -218,7 +218,7 @@ extension CurrencyFormatterTests {
         case indonesia = "id_ID"
     }
 
-    private func dollarsAndCents(from amount: Double) -> (dollars: String, cents: String) {
+    private func dollarsAndCents(from amount: Decimal) -> (dollars: String, cents: String) {
         let components = CurrencyFormatter.shared.components(from: amount)
         return ("\(components.majorUnit)\(components.decimalSeparator)", components.minorUnit)
     }
